BACK/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
tzdata==2025.2
requests==2.32.3
undetected-chromedriver
selenium-stealth
selenium
bs4
fake-useragent